Switched from u8 to usize for indexing

+++ src/blank_cell.rs
@@ -2,6 +2,6 @@
 // Blank Cell used to prioritize solving tree.
 #[derive(Debug, PartialEq)]
 pub struct BlankCell {
-	pub index: u8,
+	pub index: usize,
 	pub possible_values: Vec<u8>
 }
